Move registry from sunet::docker_compose to docker_run
These are run as 3 different containers, and not in a compose group
to simplify their running and ensure they keep running in case of
system degradation and/or cleanup scripts going berserk.

This way, we can control the startup order better, and have the
registry-auth container come out of the "backdoor" of the registry.

This way we can be independent of any images in local cache.