Add first cut of inventory script
This module with gather information from packagemanager and docker
as well as any puppet facts present on the system and put it in
a json file that can be consumed in any way you want.

Installed packages and information from docker will also be available
as facts in puppet. Inventor script is run via scripthearder every 10
minutes.

Output directory and retention time in days for collected information
is configurable.