This patch adds a simple pre-push hook
The hook will verify that the last tag matches the name of the branch
you are in. The reason for this is that we are moving to a new
development and release model for puppet-sunet where development is done
in the master branch and production machines must use a different tag
and branch than the development branch. The normal sunet-2* tag will be
used in a branch that will be called sunet-classic.

To install this hook in your local repository you need to run

  make hooks

in the root of this repository.

It is possible to bypass this hook with the --no-verify option.

If you last commit is not tagged, it will simply let the push through.