Add sunet::nftables::kvm_nat and related template
Used to create NAT and forward rules needed for the default virbr0
bridge on libvirt KVM hosts from working.

This needs to be bundled with changing <forward mode='nat'> to <forward mode='open'>
in the config file for the network. For now this is out of scope for
this change.