sunet::noc: Add a class for a noc user that can run specified tasks as root
After a meeting with NOC staff we decided that it would be a good idea
if NOC staff could ssh as a separate user that could only do predefined
tasks, hence this class. The idea is that this class can be easily
added to any cosmos ops repo and allow for a standardized way for NOC
staff to operate. The workflow is that each project gets onboarded using
this class by agreement with, and upon request from, NOC team.

The default is that noc user can reboot the server using
sunet_noc_reboot command.

Additional tasks that are allowed to run as root can be added by simply
dropping scripts in /usr/local/bin/ with the name prefix sunet_noc_

The sudoers rule for this will only allow alpha numeric characters in
the rest of the file name and you can not pass in parameters to the
file.

That means that you should only create scripts for noc user that run on
predefined input and nothing else.

NOC staff is granted access to the noc user by adding the users ssh public key
to the noc_ssh_authorized_keys hash in hiera:

noc_ssh_authorized_keys:
  kano+0DA0A7A5708FE257:
    ensure: present
    key: <public key goes here>
    name: kano@sunet.se
    type: ssh-rsa

Justification: This is safe because only the root user is allowed to put
things in /usr/local/bin. If you have modified your system to allow a
non privileged user to but arbitrary files into /usr/local/bin than that
would mean that the noc user could easily be an attack vector for
privilege escalation. However, that would only be an issue if the ssh key
of a noc user is also compromised, and in that case it is at least no
bigger issue, than if that user had been allowed root privileges as
would be the case if the noc user did not exist.