sunet-bootstrap-baas2: split up operations
Make it so we can run the tool with --install and --register flags.

This is so we can use the same tool to install things, then let puppet
modify configuration files, and then attempting to register the client
afterwards.

It also seems reasonable that we manage the "enabled" state of the
"dsmcad" service with puppet instead of inventing a custom way of doing
it in the bootstrap tool.