Add support for setting encryption password
This turned out to be more complex than we hoped for, but now we can set
both the "client" password for connecting to the BaaS service as well as
setting the encryption password without having to work around an
interactive question when running "dsmc" for the first time.

While here fix pylint warnings, mypy still needs more work.