Remove trailing newline chars from encrypt pw
The dsmc client creates password entries in TSM.KDB that includes no
newline. It is still possible to create passwords that include a newline
when creating them by hand. It is unclear how a newline would affect the
ability to decrypt things later, so for now just strip any
newline-related characters from the password before we use it.