Start using client certs for etcd3 backend
When setting up backups for the etcd backend it was nicer to supply
client certificate files instead of having to pass around a password.

For this reason I have reworked the standalone knubbis-fleetlock setup
to use client certs for all auth instead of passwords. As we regenerate
these certs each startup as well we get automatic credential rotation
and have two less values to track in our eyaml files.

This requires a bunch of changes:
* Since we need to create client certs readable by the knubbis-fleetlock
  service we now use the same uids for this process as the etcd and cert
  generation containers. The other option would have been to run the
  cert generation as root (so we could chown files to other uids) which
  seems unnecessary.
* Generate client certs with CN of "root" and "knubbis-fleetlock" to
  replace the previous password based user accounts.
* Update cfssl.json so we have a "server" profile as well as a "client"
  profile for the two kinds of certs we now generate.
* Create etcd users with `--no-password` since we now base logins on the
  CN in the client certs instead.

Some additional tweaks:
* Create /etcd-backup directory in etcd container for saving snapshots.
  Writing them to the existing data dir would cause a warning on next
  startup of etcd:
```
{"level":"warn","ts":"2023-04-21T13:30:50.913Z","caller":"etcdmain/etcd.go:446","msg":"found invalid file under data directory","filename":"snapshot.db","data-dir":"/etcd-data"}
```