dockerhost: Remove docker-compose running in docker
This patch removes the convoluted way that docker-compose is set up to
run in docker and simply installs the version that is available from the
repo.

The benefit of this is that there will not be race conditions where the
actual image that you want to run can start, but the compose image it
self has problems starting, resulting in boot loops for the actual
image.