mariadb: This patch adds a new mariadb cluster class
For now this class will run natively on the host, using distro packages
of mariadb. In the future we might move it into docker, but there are
some things regarding network namespacing that makes it hard.

To use this class, simply add it to a odd number of server. eg.:

  sunet::mariadb:
    wsrep_cluster_address: 192.121.208.250,89.45.237.4,89.46.20.160
    client_ips:
      - 192.121.208.250
      - 89.45.237.4
      - 89.46.20.160
      - 192.121.208.200
      - 89.45.237.97
      - 89.46.21.22
    id: 1

Notice that id MUST be unique to each server, while the other paramters
SHOULD be the same.

On the first host you can then run the script:
/usr/local/bin/bootstrap_cluster

Then you can simply start mariadb with systemctl on the other hosts, and
finaly restart mariadb on the first host, and you have a cluster.

Please note that mysql_backup_password MUST be set in hiera using
eyaml/edit-secrets, to the same password accross the cluster.