Handle out of order TSM.IDX entries
The bootstrap tool was crashing on a node that had been running for a
while when being run from puppet:
```
SUNET_BAAS_PASSWORD=secret1 SUNET_BAAS_ENCRYPTION_PASSWORD=secret2 /usr/local/sbin/sunet-baas2-bootstrap --register
[...]
RuntimeError: mismatch in number of found entries 0 compared to the current index we parsed: 1
```

This was caused by the code expecting (and verifying) that entries in
the TSM.IDX file would appear in an ascending order (so first entry has
index 0, second entry index 1 etc.). When inspecting TSM.IDX on the
server where the script was failing the first entry instead had index 1
while the second entry had index 0.

The reason for this appears to be that TSM will automatically rotate the
client secret over time and while the new entry retains index 0 the
new entry ends up at the end of the TSM.IDX file.

To solve this we now store the index value of each entry in our list of
parsed entries, and verify that the entry we expect to use for our
encryption secret (1) is not already taken before using it.

While here fix some comment typos.