Use better restore order for knubbis-fleetlock
Instead of moving the real data-dir out of the way, creating an empty
such directory and then restoring there we now create an empty restore
directory and mount that as the data-dir during the restore. When the
restore worked we move the original data-dir out of the way to a
date-based name and move the restore dir to the real location.

The reason for preferring restoring "first" and moving around
directories later is that we can catch things like "out if disk space"
errors without messing with the real directories which feels like a more
safe way of doing things.