Option to disable running docker in it's own ns
Needed to run containers in "host mode" or then doing other advanced operations.

At this stage running a container in non-"host mode" will require NAT rules to
nft in order to get any traffic through. That is not in scope for this commit
but we did test it and could get it to work. It might be cumbersome to run
multiple docker networks and get the NAT to work - this we did not test.

Co-authored-by: mifr <mifr@sunet.se>