BankIdP + some docker stuff (#91)
Initial commit of BankID IdP which includes  modernized syntax to work in puppet 7.x and the concept of advanced networking, needed to run containers in "host mode" or then doing other advanced operations.

At this stage running a container in non-"host mode" will require NAT rules to
nft in order to get any traffic through. That is not in scope for this commit
but we did test it and could get it to work. It might be cumbersome to run
multiple docker networks and get the NAT to work - this we did not test.