baas2: handle TSM.IDX file with multiple nodenames
Our check for duplicate index numbers in the TSM.IDX file would fail on
a machine with multiple nodenames in it. This is because the index
numbers only seem to be unique per nodename.

Now we only verify consistency for the nodename that is configured in
the configuration file instead.

Because of this potential duplication we can no longer use our static
label name either, so now we create a label based on a current timestamp
and verify it is available before using it.

Solution created with pahol/@SpaceFarmer