Only install rng-tools on older kernels
When working on a ubuntu 24.04 machine I saw this message on each cosmos
run:
```
Notice: /Stage[main]/Sunet::Tools/Package[rng-tools]/ensure: created
```

This is because rng-tools is now a virtual package installing
rng-tools-debian instead. Looking into this it seems rng-tools is not
even useful for modern kernels, from
https://wiki.archlinux.org/title/Rng-tools:
===
Note: rng-tools is not needed anymore since Kernel 5.6 because /dev/random does not block anymore.
===
... so instead of pointing to rng-tools-debian or rng-tools5 on modern
ubuntu machines just stop installing it for all machines with a kernel
of 5.6 or later.