Replace ntp with chrony on our servers
This affects Ubuntu 23.04 and up as well as Debian 12 and up.

It is unfortunate that we need to modify sunet::ntp as well, but the
problem is that ops-repos sometimes pull in sunet::ntp by hand
(sometimes to set specific configuration in cosmos-rules etc) so in
those cases it is hard to do a clean cutover between them.

For that reason sunet::ntp now does nothing on hosts matching the same
criteria as the ones where sunet::server applies sunet::chrony. This way
both can be applied to a host at the same time.