Rename "index" to "key_type"
After debugging dsmc with a command such as:
```
dsmc restore /some/file -latest -traceflags=service,verbinfo,verbdetail -tracefile=trace.out
```
... the trace debug log contains messages like these for looking up the
client password:
```
2025-01-09 08:47:31.407 [873794] [2388989760] : PasswordFile.cpp    ( 348): getPassword(): type = 0, nodeName = 'XXXXXXXXXXXX', serverName = 'SOMESERVER'
2025-01-09 08:47:31.407 [873794] [2388989760] : GSKitPasswordFile.cpp( 909): GSKitPasswordFile::readPassword: type:0 nodeName:'XXXXXXXXXXXX' serverName:'SOMESERVER' applicationType:''
```
... and log messages like these for looking up the encryption password:
```
2025-01-09 08:47:39.742 [873794] [2307909184] : PasswordFile.cpp    ( 348): getPassword(): type = 1, nodeName = 'XXXXXXXXXXXX', serverName = 'SOMESERVER'
2025-01-09 08:47:39.742 [873794] [2307909184] : GSKitPasswordFile.cpp(1404): readEncrkeyList(): Enter with node 'XXXXXXXXXXXX'.
```

So now the latest idea is that these numbers in the IDX file are not "index"
numbers referring to the position in the IDX file, but they are actually
key types (0 = client password, 1 = encryption password).

Update the code to reason about "key_type" instead of "index" to reflect
this.

While here run "black" on the code.