Update chrony config for Debian 13
On Debian 13 the default chrony config has been modified to use
leapseclist instead of leapsectz so lets do that for our Debian 13
machines as well. Once we start seeing Ubuntu machines using leapseclist
we can add a version check for Ubuntu as well.

Also Debian 13 has moved the confdir directive to the end of the config
to fix https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1073865 and
this feels like something we can do on both old and new machines.

Finally fix comment wording based on Debian 13 config so the diff is
smaller on modern machines.