Start using DNAT for knubbis-fleetlock service
Using sunet::dockerhost2 means we no longer se the real source IPs by
default because the traffic is passing through docker-proxy.

Since knubbis-fleetlock expects to ratelimit requests based on source IP
this is not very helpful. Instead statically configure the IP addresses
of the fleetlock container and use explicit DNAT rules to pass on the
traffic.