Fix logic error when checking OS version.
The if statement that looked for ubuntu24+ was only evavluated on the
KVM server OS, not the actual VM that uses the netplan config.
Instead we use a parameter to the class where you can choose to use the
new netplan routes syntax. Default to use old syntax for backwards
compability.